requirements.txt
beautifulsoup4
requests
yagooglesearch
python-dotenv
replicate
supabase
flask
